backend/pyproject.toml
[project]
name = "trend-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.112",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.8",
    "python-dotenv>=1.0",
    "SQLAlchemy>=2.0",
    "asyncpg>=0.29",
    "psycopg2-binary>=2.9",
    "alembic>=1.13",
    "httpx>=0.27",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "slowapi>=0.1.9",
    "loguru>=0.7",
    "anyio>=4.4",
]

[tool.pyright]
typeCheckingMode = "basic"
